REL-97 - Short to Integer.

+++ relayD-domain-model/src/test/java/com/relayd/attributes/ShirtsizeTest.java
@@ -26,7 +26,7 @@ public class ShirtsizeTest {
 
 	@Test
 	public void testDecodeDamenXS() {
-		Short shirtsize = Short.valueOf((short) 1);
+		Integer shirtsize = Integer.valueOf((short) 1);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -37,14 +37,14 @@ public class ShirtsizeTest {
 	public void testEncodeDamenXS() {
 		Shirtsize shirtsizeEnum = Shirtsize.DamenXS;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 1), shirtsize);
+		assertEquals(Integer.valueOf((short) 1), shirtsize);
 	}
 
 	@Test
 	public void testDecodeDamenS() {
-		Short shirtsize = Short.valueOf((short) 2);
+		Integer shirtsize = Integer.valueOf((short) 2);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -55,14 +55,14 @@ public class ShirtsizeTest {
 	public void testEncodeDamenS() {
 		Shirtsize shirtsizeEnum = Shirtsize.DamenS;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 2), shirtsize);
+		assertEquals(Integer.valueOf((short) 2), shirtsize);
 	}
 
 	@Test
 	public void testDecodeDamenM() {
-		Short shirtsize = Short.valueOf((short) 3);
+		Integer shirtsize = Integer.valueOf((short) 3);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -73,14 +73,14 @@ public class ShirtsizeTest {
 	public void testEncodeDamenM() {
 		Shirtsize shirtsizeEnum = Shirtsize.DamenM;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 3), shirtsize);
+		assertEquals(Integer.valueOf((short) 3), shirtsize);
 	}
 
 	@Test
 	public void testDecodeDamenL() {
-		Short shirtsize = Short.valueOf((short) 4);
+		Integer shirtsize = Integer.valueOf((short) 4);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -91,14 +91,14 @@ public class ShirtsizeTest {
 	public void testEncodeDamenL() {
 		Shirtsize shirtsizeEnum = Shirtsize.DamenL;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 4), shirtsize);
+		assertEquals(Integer.valueOf((short) 4), shirtsize);
 	}
 
 	@Test
 	public void testDecodeDamenXL() {
-		Short shirtsize = Short.valueOf((short) 5);
+		Integer shirtsize = Integer.valueOf((short) 5);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -109,14 +109,14 @@ public class ShirtsizeTest {
 	public void testEncodeDamenXL() {
 		Shirtsize shirtsizeEnum = Shirtsize.DamenXL;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 5), shirtsize);
+		assertEquals(Integer.valueOf((short) 5), shirtsize);
 	}
 
 	@Test
 	public void testDecodeHerrenS() {
-		Short shirtsize = Short.valueOf((short) 6);
+		Integer shirtsize = Integer.valueOf((short) 6);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -127,14 +127,14 @@ public class ShirtsizeTest {
 	public void testEncodeHerrenS() {
 		Shirtsize shirtsizeEnum = Shirtsize.HerrenS;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 6), shirtsize);
+		assertEquals(Integer.valueOf((short) 6), shirtsize);
 	}
 
 	@Test
 	public void testDecodeHerrenM() {
-		Short shirtsize = Short.valueOf((short) 7);
+		Integer shirtsize = Integer.valueOf((short) 7);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -145,14 +145,14 @@ public class ShirtsizeTest {
 	public void testEncodeHerrenM() {
 		Shirtsize shirtsizeEnum = Shirtsize.HerrenM;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 7), shirtsize);
+		assertEquals(Integer.valueOf((short) 7), shirtsize);
 	}
 
 	@Test
 	public void testDecodeHerrenL() {
-		Short shirtsize = Short.valueOf((short) 8);
+		Integer shirtsize = Integer.valueOf((short) 8);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -163,14 +163,14 @@ public class ShirtsizeTest {
 	public void testEncodeHerrenL() {
 		Shirtsize shirtsizeEnum = Shirtsize.HerrenL;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 8), shirtsize);
+		assertEquals(Integer.valueOf((short) 8), shirtsize);
 	}
 
 	@Test
 	public void testDecodeHerrenXL() {
-		Short shirtsize = Short.valueOf((short) 9);
+		Integer shirtsize = Integer.valueOf((short) 9);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -181,14 +181,14 @@ public class ShirtsizeTest {
 	public void testEncodeHerrenXL() {
 		Shirtsize shirtsizeEnum = Shirtsize.HerrenXL;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 9), shirtsize);
+		assertEquals(Integer.valueOf((short) 9), shirtsize);
 	}
 
 	@Test
 	public void testDecodeHerrenXXL() {
-		Short shirtsize = Short.valueOf((short) 10);
+		Integer shirtsize = Integer.valueOf((short) 10);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
 
@@ -199,14 +199,14 @@ public class ShirtsizeTest {
 	public void testEncodeHerrenXXL() {
 		Shirtsize shirtsizeEnum = Shirtsize.HerrenXXL;
 
-		Short shirtsize = Shirtsize.encode(shirtsizeEnum);
+		Integer shirtsize = Shirtsize.encode(shirtsizeEnum);
 
-		assertEquals(Short.valueOf((short) 10), shirtsize);
+		assertEquals(Integer.valueOf((short) 10), shirtsize);
 	}
 
 	@Test
 	public void testDecodeInvalidValue() {
-		Short shirtsize = Short.valueOf((short) 99);
+		Integer shirtsize = Integer.valueOf((short) 99);
 
 		Shirtsize shirtsizeEnum = Shirtsize.decode(shirtsize);
